set xEnemies

+++ mainwindow.h
@@ -42,7 +42,7 @@ private slots:
 
     void on_pushButton_3_clicked();
 
-    void on_health_valueChanged(int value);
+    void on_health_valueChanged(float value);
 
 private:
     Ui::MainWindow *ui;
